Fixes the UAL login with Wombat
The ual-wombat library has several transitive dependencies
that require access to certain objects on `window` that are
usually provided by Webpack I guess. It seems nuxt.js is not
providing this.

To circumvent this the code now adapts to it. I made sure client
relevant code is only executed on the client and also provided
`window.global` (required by readable-stream) as well as
`window.process.env (required by util).

This made login and signing a transaction (I didn't test broadcasting
it) possible on my machine with the Wombat Chrome extension.