Uses official Wombat UAL package, adds error handling
Switches to the official Wombat UAL package which was just released in
version 1.0.0 to fix the broken build with Nuxt. The transitive
dependency on Scatter.js that was at fault was removed.

This allows some simplifications (`window.global = window` e.g.).

Also adds some error handling to the `onMounted` hook to make errors
better visible (rejected promises in `onMounted` are not automatically
handled).